requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asgiref==3.7.2
astroid==3.0.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==4.2.0
cachetools==5.3.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
contrib==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==41.0.4
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.2.5
django-allauth==0.57.0
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-configurations==2.5
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-etc==1.4.0
django-filter==23.5
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==1.2.0
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==7.0.1
inflection==0.5.1
ipython==8.14.0
isort==5.12.0
itypes==1.2.0
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.14.0
keyring==24.3.0
kiwisolver==1.4.5
kombu==5.3.5
libclang==16.0.6
magic-filter==1.0.12
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
ml-dtypes==0.2.0
more-itertools==10.1.0
multidict==6.0.4
numpy==1.26.1
oauthlib==3.2.2
openapi-codec==1.3.2
opt-einsum==3.3.0
packaging==23.2
paramiko==3.4.0
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.31
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.38
protobuf==4.24.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
PyJWT==2.8.0
pylint==3.0.1
PyNaCl==1.5.0
pyparsing==3.1.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.15.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-environ==0.4.54
python-telegram-bot==20.7
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.2
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.16.2
rsa==4.9
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
storage==0.0.4.3
telebot==0.0.5
telegram==0.0.1
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
tomlkit==0.12.1
traitlets==5.9.0
transliterate==1.10.2
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.4
vine==5.1.0
virtualenv==20.24.4
wcwidth==0.2.6
Werkzeug==3.0.0
wrapt==1.14.1
yarl==1.9.2
zipp==3.17.0